+++ app/src/main/java/edu/ucne/jugadorestictactoe/data/local/jugadores/Dao/JugadorDao.kt
@@ -1,10 +1,10 @@
-package edu.ucne.jugadorestictactoe.data.local.Dao
+package edu.ucne.jugadorestictactoe.data.local.jugadores.Dao
 
 import androidx.room.Dao
 import androidx.room.Delete
 import androidx.room.Query
 import androidx.room.Upsert
-import edu.ucne.jugadorestictactoe.data.local.Entities.JugadorEntity
+import edu.ucne.jugadorestictactoe.data.local.jugadores.Entities.JugadorEntity
 import kotlinx.coroutines.flow.Flow
 
 @Dao
