+++ app/src/main/java/edu/ucne/jugadorestictactoe/data/local/mappers/mappers.kt
@@ -1,0 +1,37 @@
+package edu.ucne.jugadorestictactoe.data.local.mappers
+
+import edu.ucne.jugadorestictactoe.data.local.jugadores.Entities.JugadorEntity
+import edu.ucne.jugadorestictactoe.data.local.partidas.Entity.PartidaEntity
+import edu.ucne.jugadorestictactoe.domain.model.Jugador
+import edu.ucne.jugadorestictactoe.domain.model.Partida
+
+
+fun JugadorEntity.toDomain() = Jugador(
+        id = jugadorId ?: 0,
+        nombre = nombres,
+        partidas = partidas
+    )
+
+    fun Jugador.toEntity() = JugadorEntity(
+        jugadorId = id,
+        nombres = nombre,
+        partidas = partidas
+    )
+
+    fun PartidaEntity.toDomain() = Partida(
+        id = partidaId ?: 0,
+        fecha = fecha,
+        jugador1 = jugador1Id,
+        jugador2 = jugador2Id,
+        ganadorId = ganadorId,
+        esFinalizada = esFinalizada
+    )
+
+fun Partida.toEntity() = PartidaEntity(
+        partidaId = id,
+        fecha = fecha,
+        jugador1Id = jugador1,
+        jugador2,
+        ganadorId = ganadorId,
+        esFinalizada = esFinalizada
+    )
