+++ app/src/main/java/edu/ucne/jugadorestictactoe/domain/useCase/JugadoresUseCase/ValidarJugadorUseCase.kt
@@ -1,4 +1,4 @@
-package edu.ucne.jugadorestictactoe.domain.useCase
+package edu.ucne.jugadorestictactoe.domain.useCase.JugadoresUseCase
 
 import edu.ucne.jugadorestictactoe.domain.model.Jugador
 import edu.ucne.jugadorestictactoe.domain.repository.JugadorRepository
